# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CryptoStake Casino API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kyc:
    get:
      operationId: getKyc
      tags: [kyc]
      summary: Get KYC status for current user
      responses:
        "200":
          description: KYC data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
    post:
      operationId: submitKyc
      tags: [kyc]
      summary: Submit KYC information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "201":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"

  # ─── Games ───────────────────────────────────────────────────────────────────
  /games/dice:
    post:
      operationId: playDice
      tags: [games]
      summary: Play a dice game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiceInput"
      responses:
        "200":
          description: Dice result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiceResult"
  /games/coinflip:
    post:
      operationId: playCoinflip
      tags: [games]
      summary: Play a coin flip game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoinflipInput"
      responses:
        "200":
          description: Coin flip result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinflipResult"
  /games/blackjack/start:
    post:
      operationId: startBlackjack
      tags: [games]
      summary: Start a new blackjack hand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlackjackStart"
      responses:
        "200":
          description: Blackjack hand started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlackjackState"
  /games/blackjack/{sessionId}/action:
    post:
      operationId: blackjackAction
      tags: [games]
      summary: Perform a blackjack action (hit, stand, double)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlackjackAction"
      responses:
        "200":
          description: Updated blackjack state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlackjackState"
  /games/slots:
    post:
      operationId: playSlots
      tags: [games]
      summary: Spin the slot machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotsInput"
      responses:
        "200":
          description: Slots result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsResult"
  /games/history:
    get:
      operationId: getGameHistory
      tags: [games]
      summary: Get recent game history for current user
      parameters:
        - name: game
          in: query
          schema:
            type: string
            enum: [dice, coinflip, blackjack, slots]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Game history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameHistoryItem"

  # ─── Admin ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    KycRecord:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        status:      { type: string, enum: [pending, approved, rejected] }
        fullName:    { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        country:     { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    KycInput:
      type: object
      required: [fullName, dateOfBirth, country]
      properties:
        fullName:    { type: string }
        dateOfBirth: { type: string }
        country:     { type: string }
        idDocumentUrl: { type: string }
        selfieUrl:   { type: string }
...
    DiceInput:
      type: object
      required: [wager, currency, winChance, isOver]
      properties:
        wager:      { type: string }
        currency:   { type: string, enum: [ETH, BTC, SOL, USDC] }
        winChance:  { type: number }
        isOver:     { type: boolean }
    DiceResult:
      type: object
      required: [roll, winChance, isOver, won, payout, newBalance, transactionId]
      properties:
        roll:          { type: number }
        winChance:     { type: number }
        isOver:        { type: boolean }
        won:           { type: boolean }
        payout:        { type: string }
        newBalance:    { type: string }
        transactionId: { type: integer }
    CoinflipInput:
      type: object
      required: [wager, currency, choice]
      properties:
        wager:    { type: string }
        currency: { type: string, enum: [ETH, BTC, SOL, USDC] }
        choice:   { type: string, enum: [heads, tails] }
    CoinflipResult:
      type: object
      required: [outcome, won, payout, newBalance, transactionId]
      properties:
        outcome:       { type: string, enum: [heads, tails] }
        won:           { type: boolean }
        payout:        { type: string }
        newBalance:    { type: string }
        transactionId: { type: integer }
    BlackjackStart:
      type: object
      required: [wager, currency]
      properties:
        wager:    { type: string }
        currency: { type: string, enum: [ETH, BTC, SOL, USDC] }
    BlackjackState:
      type: object
      required: [sessionId, playerCards, dealerCards, playerScore, dealerScore, status, wager, currency]
      properties:
        sessionId:    { type: string }
        playerCards:  { type: array, items: { type: string } }
        dealerCards:  { type: array, items: { type: string } }
        playerScore:  { type: integer }
        dealerScore:  { type: integer }
        status:       { type: string, enum: [playing, player_bust, dealer_bust, player_win, dealer_win, push, blackjack] }
        wager:        { type: string }
        currency:     { type: string }
        payout:       { type: ["string", "null"] }
        transactionId: { type: ["integer", "null"] }
    BlackjackAction:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [hit, stand, double] }
    SlotsInput:
      type: object
      required: [wager, currency]
      properties:
        wager:    { type: string }
        currency: { type: string, enum: [ETH, BTC, SOL, USDC] }
    SlotsResult:
      type: object
      required: [reels, won, multiplier, payout, newBalance, transactionId]
      properties:
        reels:         { type: array, items: { type: string }, minItems: 3, maxItems: 3 }
        won:           { type: boolean }
        multiplier:    { type: number }
        payout:        { type: string }
        newBalance:    { type: string }
        transactionId: { type: integer }
    GameHistoryItem:
      type: object
      required: [id, gameType, wager, currency, won, payout, createdAt]
      properties:
        id:          { type: integer }
        gameType:    { type: string }
        wager:       { type: string }
        currency:    { type: string }
        won:         { type: boolean }
        payout:      { type: string }
        createdAt:   { type: string }

    # ─── Admin ───────────────────────────────────────────────────────────────
...