# lib/api-zod/src/generated/types/coinflipResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CryptoStake Casino API
 * OpenAPI spec version: 0.1.0
 */
import type { CoinflipResultOutcome } from './coinflipResultOutcome';

export interface CoinflipResult {
  outcome: CoinflipResultOutcome;
  won: boolean;
  payout: string;
  newBalance: string;
  transactionId: number;
}